# backend/openapi.yaml
openapi: 3.0.0
info:
  title: RentHub API
  description: Rental Platform API Documentation
  version: 1.0.0
  contact:
    name: API Support
    email: api@renthub.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development
  - url: https://api.renthub.com/v1
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Property:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        type:
          type: string
          enum: [apartment, house, villa, studio]
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        address:
          type: string
        city:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [available, booked, maintenance]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
paths:
  /properties:
    get:
      tags:
        - Properties
      summary: List all properties
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 15
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: min_price
          schema:
            type: number
        - in: query
          name: max_price
          schema:
            type: number
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                      total:
                        type: integer
                      per_page:
                        type: integer

    post:
      tags:
        - Properties
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Property'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...